pyproject.toml
[project]
name = "ai-orchestrator"
version = "0.1.0"
description = "Telegram-based AI orchestrator for Mac Mini M4 Pro remote control"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[all]>=22.0",
    "psutil>=6.0",
    "anthropic>=0.40",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["orchestrator"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
